refactor: `useRender()` (#273)
* refactor: `useRender()`

* chore: review comment
